requirements.txt
Flask==1.1.1
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.4.0
flask-marshmallow==0.10.1
marshmallow-sqlalchemy==0.19.0
marshmallow==2.18.0
apache-airflow==1.10.5
pyspark==2.4.4
requests==2.22.0
vaderSentiment==3.2.1
python-geohash==0.8.5
psycopg2-binary==2.8.3
GeoAlchemy2==0.6.3